backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
pydantic-settings==2.9.1
certifi==2025.4.26
click==8.1.8
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
pydantic==2.11.4
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
ujson==5.10.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
yt-dlp==2024.8.1
cryptography==3.4
passlib==1.7.4
python-jose==3.4.0
python-multipart==0.0.20
# playwright==1.41.2
# tenacity==8.2.3
greenlet==3.0.3